updated list view with no items view if the list of items is empty

+++ TodoList/Views/ListView.swift
@@ -12,23 +12,31 @@ struct ListView: View {
     @EnvironmentObject var listViewModel: ListViewModel
     
     var body: some View {
-        List {
-            ForEach(listViewModel.items) { item in
-                ListRowView(item: item)
-                    .onTapGesture {
-                        withAnimation(.linear) {
-                            listViewModel.updateItem(item: item)
-                        }
+        ZStack {
+            if listViewModel.items.isEmpty {
+                NoItemsView()
+                    .transition(AnyTransition.opacity.animation(.easeIn))
+            }
+            else {
+                List {
+                    ForEach(listViewModel.items) { item in
+                        ListRowView(item: item)
+                            .onTapGesture {
+                                withAnimation(.linear) {
+                                    listViewModel.updateItem(item: item)
+                                }
+                            }
                     }
-            }
-            .onDelete { indexSet in
-                listViewModel.deleteItem(indexSet: indexSet)
-            }
-            .onMove { indexSet, to in
-                listViewModel.moveItem(from: indexSet, to: to)
+                    .onDelete { indexSet in
+                        listViewModel.deleteItem(indexSet: indexSet)
+                    }
+                    .onMove { indexSet, to in
+                        listViewModel.moveItem(from: indexSet, to: to)
+                    }
+                }
+                .listStyle(PlainListStyle())
             }
         }
-        .listStyle(PlainListStyle())
         .navigationTitle("Todo List 📝")
         .navigationBarItems(
             leading: EditButton(),
